minor storage optimization
I think this should save a slot, since it should get packed now.
Commit switches from using a uint256 to a uint32 for storing a timestamp (consistent with our use elsewhere) and reorders a struct definition.

+++ src/contracts/middleware/RegistryBase.sol
@@ -304,7 +304,7 @@ abstract contract RegistryBase is
         registry[msg.sender].serveUntil = repository.serviceManager().latestTime();
         // committing to not signing off on any more middleware tasks
         registry[msg.sender].active = IQuorumRegistry.Active.INACTIVE;
-        registry[msg.sender].deregisterTime = block.timestamp;
+        registry[msg.sender].deregisterTime = uint32(block.timestamp);
 
         // gas saving by caching length here
         uint256 pubkeyHashToStakeHistoryLengthMinusOne = pubkeyHashToStakeHistory[pubkeyHash].length - 1;
